+++ index.js
@@ -1,6 +1,5 @@
 var stripANSI = require('strip-ansi');
 var path = require('path');
-var objectAssign = require('object-assign');
 var os = require('os');
 var notifier = require('node-notifier');
 
@@ -13,43 +12,70 @@ function WebpackNotifierPlugin(options) {
 }
 module.exports = WebpackNotifierPlugin;
 
-WebpackNotifierPlugin.prototype.compileMessage = function (stats) {
-  function findFirstDFS(compilation, key) {
-    var match = compilation[key][0];
+function findFirstDFS(compilation, key) {
+  var match = compilation[key][0];
+  if (match) {
+    return match;
+  }
+
+  var children = compilation.children;
+  for (var i = 0; i < children.length; i += 1) {
+    match = findFirstDFS(children[i], key);
     if (match) {
       return match;
     }
-
-    var children = compilation.children;
-    for (var i = 0; i < children.length; i += 1) {
-      match = findFirstDFS(children[i], key);
-      if (match) {
-        return match;
-      }
-    }
   }
+}
 
+WebpackNotifierPlugin.prototype.compileEndOptions = function (stats) {
   if (this.isFirstBuild) {
     this.isFirstBuild = false;
 
     if (this.options.skipFirstNotification) {
-      return;
+      return {};
     }
+  }
+
+  var imageFromOptions = ('contentImage' in this.options)
+    ? this.options.contentImage
+    : DEFAULT_LOGO;
+
+  var successImage = '';
+  var warningsImage = '';
+  var errorsImage = '';
+  if (typeof imageFromOptions === 'object') {
+    successImage = imageFromOptions.success;
+    warningsImage = imageFromOptions.warning;
+    errorsImage = imageFromOptions.error;
+  } else {
+    successImage = imageFromOptions;
+    warningsImage = imageFromOptions;
+    errorsImage = imageFromOptions;
   }
 
   var hasEmoji = this.options.emoji;
   var error;
+  var contentImage;
+  var status;
   if (this.hasErrors(stats)) {
     error = findFirstDFS(stats.compilation, 'errors');
+    contentImage = errorsImage;
+    status = 'error';
   } else if (this.options.onlyOnError) {
-    return;
+    return {};
   } else if (this.hasWarnings(stats) && !this.options.excludeWarnings) {
     error = findFirstDFS(stats.compilation, 'warnings');
+    contentImage = warningsImage;
+    status = 'warning';
   } else if (!this.lastBuildSucceeded || this.options.alwaysNotify) {
     this.lastBuildSucceeded = true;
-    return (hasEmoji ? '✅ ' : '') + 'Build successful';
+    return {
+      message: (hasEmoji ? '✅ ' : '') + 'Build successful',
+      contentImage: successImage,
+      status: 'success'
+    };
   } else {
-    return;
+    return {};
   }
 
   this.lastBuildSucceeded = false;
@@ -67,7 +93,11 @@ WebpackNotifierPlugin.prototype.compileMessage = function (stats) {
     message = (hasEmoji ? '⚠️ ' : '') + 'Warning: ' + message + error.message.toString();
   }
 
-  return stripANSI(message);
+  return {
+    message: stripANSI(message),
+    contentImage: contentImage,
+    status: status
+  };
 };
 
 WebpackNotifierPlugin.prototype.hasErrors = function (stats) {
@@ -81,22 +111,31 @@ WebpackNotifierPlugin.prototype.hasWarnings = function (stats) {
 };
 
 WebpackNotifierPlugin.prototype.compilationDone = function (stats) {
-  var msg = this.compileMessage(stats);
-  if (msg) {
-    var contentImage = ('contentImage' in this.options)
-      ? this.options.contentImage : DEFAULT_LOGO;
-
-    var title = this.options.title;
+  var { message, contentImage, status } = this.compileEndOptions(stats);
+  if (message) {
+    var title = this.options.title ? this.options.title : 'Webpack';
     if (typeof title === 'function') {
-      title = title({ msg: msg });
+      title = title({
+        msg: message, // compatibility with v1.11.0
+        message: message,
+        status: status
+      });
     }
 
-    notifier.notify(objectAssign({
-      title: 'Webpack',
-      message: msg,
-      contentImage: contentImage,
-      icon: (os.platform() === 'win32' || os.platform() === 'linux') ? contentImage : undefined
-    }, this.options, { title: title }));
+    var icon = (os.platform() === 'win32' || os.platform() === 'linux')
+      ? contentImage
+      : undefined;
+
+    notifier.notify(Object.assign(
+      {},
+      this.options,
+      {
+        title,
+        message,
+        contentImage,
+        icon
+      }
+    ));
   }
 };